+++ README.md
@@ -30,7 +30,7 @@ Here are the files you need to create :
 Here are the files you need to update :
 
 - src/App.js (declare new route to access the api form)
-- src/pages/AdminEnrollmentList.js (declare API Labels in enrollment list view)
+- src/lib/api.js (declare API Labels in TARGET_API_LABELS)
 - (optional) src/lib/utils.js (L~38) (declare valid naf code)
 
 ## License
