+++ src/lib/api.js
@@ -1,0 +1,23 @@
+export const TARGET_API_LABELS = {
+  api_particulier: 'API Particulier',
+  franceconnect: 'FranceConnect',
+  api_droits_cnam: 'API Droits CNAM',
+  api_impot_particulier: 'Impôt particulier',
+  api_impot_particulier_step2: 'Impôt particulier 2/2',
+  api_entreprise: 'API Entreprise',
+  preuve_covoiturage: 'Registre de Preuve de Covoiturage',
+};
+
+export const TARGET_API_WITH_ENROLLMENTS_IN_PRODUCTION_ENV = [
+  'franceconnect',
+  'api_particulier',
+  'api_entreprise',
+  'api_impot_particulier',
+];
+
+export const API_ICONS = {
+  franceconnect: 'logo-fc.png',
+  api_droits_cnam: 'logo-cnam.jpg',
+  api_impot_particulier: 'logo-dgfip.png',
+  api_impot_particulier_step2: 'logo-dgfip.png',
+};
