+++ src/pages/ApiDroitsCnam.js
@@ -1,5 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+
+import { API_ICONS } from '../lib/api';
+
 import Form from '../components/Form';
 import Nav from '../components/Nav';
 import TextSection from '../components/form-sections/TextSection';
@@ -44,6 +47,10 @@ const ApiDroitsCnam = ({
 }) => (
   <div className="dashboard">
     <Nav
+      logo={{
+        src: `/images/${API_ICONS.api_droits_cnam}`,
+        alt: 'Caisse Nationale d’Assurance Maladie',
+      }}
       navLinks={[
         { id: 'description', label: 'Description' },
         { id: 'cadre-juridique', label: 'Cadre juridique' },
